Added Pan UI. Fixed Zoom/Pan glitches.

+++ DepthView-src/sidebysiderender.h
@@ -11,7 +11,7 @@ public:
 
     SideBySideRender();
 
-    QImage draw(QImage imgL, QImage imgR, float panX, float panY, int finalwidth = 0, int finalheight = 0, float zoom = 0);
+    QImage draw(QImage imgL, QImage imgR, int panX, int panY, int finalwidth = 0, int finalheight = 0, float zoom = 0);
 };
 
 #endif // SIDEBYSIDERENDER_H
